+++ CHANGELOG.md
@@ -35,6 +35,7 @@ FlatLaf Change Log
 - Native Windows libraries: Fixed crash when running in Java 8 and newer Java
   version is installed in `PATH` environment variable and using class
   `SystemInfo` before AWT initialization. (issue #673)
+- ComboBox: Fixed search in item list for text with spaces. (issue #691)
 - FormattedTextField: On Linux, fixed `IllegalArgumentException: Invalid
   location` if `JFormattedTextField.setDocument()` is invoked in a focus gained
   listener on that formatted text field. (issue #698)